tmp - fix tests, move sequence_encoder_registry, fix some imports.
