+++ ludwig/encoders/sequence_encoders.py
@@ -21,14 +21,13 @@ from torch import nn
 
 from ludwig.constants import AUDIO, SEQUENCE, TEXT, TIMESERIES
 from ludwig.encoders.base import Encoder
-from ludwig.encoders.registry import register_encoder
+from ludwig.encoders.registry import register_encoder, register_sequence_encoder
 from ludwig.modules.attention_modules import TransformerStack
 from ludwig.modules.convolutional_modules import Conv1DStack, ParallelConv1D, ParallelConv1DStack
 from ludwig.modules.embedding_modules import EmbedSequence, TokenAndPositionEmbedding
 from ludwig.modules.fully_connected_modules import FCStack
 from ludwig.modules.recurrent_modules import RecurrentStack
 from ludwig.modules.reduction_modules import SequenceReducer
-from ludwig.schema.combiners.utils import register_sequence_encoder
 
 logger = logging.getLogger(__name__)
