+++ ludwig/schema/combiners/__init__.py
@@ -1,8 +1,8 @@
 # Pull configs here for convenient imports in other modules:
-from comparator import ComparatorCombinerConfig  # noqa: F401
-from concat import ConcatCombinerConfig  # noqa: F401
-from sequence import SequenceCombinerConfig  # noqa: F401
-from sequence_concat import SequenceConcatCombinerConfig  # noqa: F401
-from tab_transformer import TabTransformerCombinerConfig  # noqa: F401
-from tabnet import TabNetCombinerConfig  # noqa: F401
-from transformer import TransformerCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.comparator import ComparatorCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.concat import ConcatCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.sequence import SequenceCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.sequence_concat import SequenceConcatCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.tab_transformer import TabTransformerCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.tabnet import TabNetCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.transformer import TransformerCombinerConfig  # noqa: F401
